workers/asr/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
transformers==4.57.1
huggingface_hub==0.34.0
torch==2.6.0
torchaudio==2.6.0
numpy==1.26.4
soundfile==0.12.1
sentencepiece==0.1.99
hydra-core==1.3.2
omegaconf==2.3.0
pyannote.audio==3.1.1

